third_party/blink/web_tests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252_include=xml-expected.txt

This is a testharness.js-based test.
[FAIL] <?xml-stylesheet?> (CSS)
  assert_equals: expected "\\"%C3%A5\\"" but got "\\"%E5\\""
Harness: the test ran to completion.
